+++ src/store/index.ts
@@ -44,6 +44,7 @@ export {
   setResourceAdapter,
   loadResources,
   uploadResource,
+  addUrlResource,
   updateResource,
   deleteResource,
   getResourceBlob,